DevSecOps_Agents/requirements.txt
# Core AI and ML libraries
groq==0.4.2
langchain==0.0.350
langchain-groq==0.0.1
langchain-community==0.0.10

# API and HTTP libraries
requests==2.31.0
aiohttp==3.9.1
fastapi==0.104.1
uvicorn==0.24.0

# Data processing and analysis
pandas==2.1.4
# Using pre-compiled numpy to avoid build issues
numpy==1.24.3
python-dateutil==2.8.2

# GitHub API integration
PyGithub==1.59.1
github3.py==4.0.1

# Kubernetes integration
kubernetes==28.1.0

# Docker integration
docker==6.1.3

# Monitoring and alerting
prometheus-client==0.19.0
psutil==5.9.6

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Password manager integration
keyring==24.3.0
cryptography==41.0.7

# Logging and debugging
loguru==0.7.2
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security scanning
bandit==1.7.5
safety==2.3.5

# Database (for storing monitoring data)
sqlalchemy==2.0.23
alembic==1.13.1

# Async support
asyncio-mqtt==0.16.1
websockets==12.0 